requirement.txt
streamlit
numpy
librosa
scikit-learn
pandas
pickle-mixin
python-dotenv